Use Bank AccountID and Ensure Correct Acct. Select
The bank accounts in Xero's default demo company do not
have account codes so we should really use the built-in
AccountID (GUID) to test accounts because we know it is
always there. 

The '.select' methods were occasionally not working because
the inner-parameter of the block was named the same as the
outer. Specifically use a different inner-parameter for
the block to resolve this issue.