Add datetime_utc type to http_proxy
This allows fields of type datetime_utc to be used in where clauses.
The datetime_utc type was added in commit
cdae18e1694e3ce6aa9d333f3443f6d4591ad283 but was never added to
base_model_http_proxy, so you get empty where conditions when you
try to filter on fields of that type. This commit fixes that.