Update fetching_bank_transaction_tests to test for returned attributes
Possible the API has changed what it returns from detailed/non-detailed queries, so the tests failed. Have updated and included a key check on the returned object more accurately.
Also, passing an item code when creating a new transaction fails as it cannot find this item, so have removed it.