Coerce decimals when calculating line amount
This is particularly important because String * Numeric has valid (but
very surprising) results. If quantity is supplied as a string, and
unit_amount as a numeric, line_amount will be successfully but
incorrectly calculated. e.g.

"1" * 12 # => "111111111111"

This subsequently gets coerced into a BigDecimal before being returned.

Ideally, type coercion or checking should happen at the interface of
the class, but this is the safest change to make at the moment.