Conditionally validate line item description
When item_code is given, Xero API does not require the description to be set, 
as it will pull the description from given item. Without this change it's not possible
to create an invoice without setting the description, as it will raise the 
"Line items must all be valid" error.